Solution to Calculate Employee Bonus

+++ Calculate Employee Bonus@@67877d61dc107486ad5c274f/code.c
@@ -2,7 +2,7 @@
 struct salary{
     int id;
     char str[50];
-    int salary;
+    float salary;
 };
 
 int main(){
@@ -10,7 +10,7 @@ int main(){
     scanf("%d",&n);
     struct salary sal[n];
     for(int i=0;i<n;i++){
-        scanf("%d %s %d",&sal[i].id,sal[i].str,&sal[i].salary);
+        scanf("%d %s %f",&sal[i].id,sal[i].str,&sal[i].salary);
     }
     float bonus;
     for(int i=0;i<n;i++){
